Initial Commit (Binary Search Approach).
In this problem, there are multiple power stations and in one of the station the circuit got fused and we have to find that station or bulb where the fault occur. The circuit is shown in the form of a binary array.
.
Time Complexity: O(long), where 'n' is the number of bulbs/power stations.