Initial Commit.
In this problem, we have to implement the 'atoi()' function. The 'atoi()' function takes the string as a parameter and if the string can be converted into an integer, then it will return its integer value and it will also take care of the sign of the integer.
.
Time Complexity: O(n), where n is the length of the string.